# backend/internal/assets/openapi.yaml
openapi: "3.0.3"
info:
  title: Statify API
  description: API for managing projects and static site deployments.
  version: "1.0.0"
servers:
  - url: /api/v1
paths:
...
  /projects/{project_id}/deployments:
    get:
      tags:
        - Deployments
      summary: Get Deployment History
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      tags:
        - Deployments
      summary: Create Deployment (Upload Zip)
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /projects/{project_id}/deployments/{id}:
    get:
      tags:
        - Deployments
      summary: Get Single Deployment Status
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
    delete:
      tags:
        - Deployments
      summary: Delete Deployment
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Deployment:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        status:
          type: string
          enum: [QUEUED, UPLOADED, READY, FAILED, DELETED]
        output_prefix:
          type: string
          nullable: true
        source_zip_object_key:
          type: string
          nullable: true
        validation_error:
          type: string
          nullable: true
        create_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true